integrate each collection to shared-collection
