+++ authentication/src/main/kotlin/com/kohan/authentication/repository/UserRepository.kt
@@ -1,9 +1,9 @@
 package com.kohan.authentication.repository
 
-import com.kohan.authentication.collection.UserCollection
 import org.bson.types.ObjectId
 import org.springframework.data.mongodb.repository.MongoRepository
 import org.springframework.data.mongodb.repository.Query
+import com.kohan.shared.collection.user.UserCollection
 
 interface UserRepository : MongoRepository<UserCollection, ObjectId> {
     fun findByEmail(email: String): UserCollection?
