+++ authentication/src/main/kotlin/com/kohan/authentication/util/TokenGenerator.kt
@@ -1,7 +1,7 @@
 package com.kohan.authentication.util
 
-import com.kohan.authentication.collection.item.TokenInfo
 import com.kohan.shared.armeria.converter.request.result.AccessDeviceInfo
+import com.kohan.shared.collection.user.item.TokenInfo
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
 import java.security.SecureRandom
@@ -23,6 +23,7 @@ class TokenGenerator(
             token = Base64.getEncoder().encodeToString(randomArray),
             accessDeviceInfo = accessDeviceInfo,
             expirationMillis = expiration,
+            issuanceDate = issuanceDate,
         )
     }
 }
