+++ message-rest/src/main/kotlin/com/kohan/message/rest/MessageRestApplication.kt
@@ -1,11 +1,12 @@
 package com.kohan.message.rest
 
+import com.kohan.shared.spring.mongo.config.EmbeddedMongoConfig
 import com.kohan.shared.spring.mongo.config.MongoConfig
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import org.springframework.context.annotation.Import
 
-@Import(MongoConfig::class)
+@Import(MongoConfig::class, EmbeddedMongoConfig::class)
 @SpringBootApplication
 class MessageRestApplication
