--- message-rest/src/main/kotlin/com/kohan/message/rest/enum/collection/chatRoom/ChatRoomType.kt
@@ -1,6 +1,0 @@
-package com.kohan.message.rest.enum.collection.chatRoom
-
-enum class ChatRoomType {
-    GROUP,
-    PERSONAL,
-}
