--- message-rest/src/main/kotlin/com/kohan/message/rest/enum/collection/message/item/FileType.kt
@@ -1,6 +1,0 @@
-package com.kohan.message.rest.enum.collection.message.item
-
-enum class FileType {
-    IMAGE,
-    FILE,
-}
