--- message-rest/src/main/kotlin/com/kohan/message/rest/enum/collection/message/item/ReactionType.kt
@@ -1,8 +1,0 @@
-package com.kohan.message.rest.enum.collection.message.item
-
-enum class ReactionType {
-    LIKE,
-    HEART,
-    THUMB,
-    SAD,
-}
