+++ settings.gradle.kts
@@ -10,3 +10,4 @@ include("file")
 include("message-websocket")
 include("push")
 include("proto")
+include("shared-collection")
