+++ shared-collection/build.gradle.kts
@@ -1,0 +1,4 @@
+dependencies {
+    implementation(project(":shared-armeria"))
+    implementation(libs.spring.boot.starter.data.mongodb)
+}
