+++ shared-collection/src/main/kotlin/com/kohan/shared/enum/chatRoom/ChatRoomType.kt
@@ -1,0 +1,6 @@
+package com.kohan.shared.enum.chatRoom
+
+enum class ChatRoomType {
+    GROUP,
+    PERSONAL,
+}
