+++ shared-collection/src/main/kotlin/com/kohan/shared/enum/message/item/FileType.kt
@@ -1,0 +1,6 @@
+package com.kohan.shared.enum.message.item
+
+enum class FileType {
+    IMAGE,
+    FILE,
+}
