+++ shared-collection/src/main/kotlin/com/kohan/shared/enum/message/item/ReactionType.kt
@@ -1,0 +1,8 @@
+package com.kohan.shared.enum.message.item
+
+enum class ReactionType {
+    LIKE,
+    HEART,
+    THUMB,
+    SAD,
+}
