+++ shared-spring/build.gradle.kts
@@ -23,11 +23,13 @@ dependencies {
                     .name,
         )
     }
+    api(project(":shared-collection"))
     api(libs.kotlin.allopen)
     api(libs.armeria.spring.boot3.starter)
     api(libs.spring.boot.starter.security)
     api(libs.spring.boot.starter.validation)
     api(libs.spring.boot.starter.data.mongodb)
+    api(libs.spring.kafka)
     api(libs.bcprov.jdk18on)
     api(libs.spring.embed.mongo)
