+++ shared-spring/src/main/kotlin/com/kohan/shared/spring/kafka/config/KafkaConsumerConfig.kt
@@ -1,0 +1,33 @@
+package com.kohan.shared.spring.kafka.config
+
+import org.apache.kafka.clients.consumer.ConsumerConfig
+import org.apache.kafka.common.serialization.StringDeserializer
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.kafka.annotation.EnableKafka
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory
+import org.springframework.kafka.core.DefaultKafkaConsumerFactory
+
+@EnableKafka
+@Configuration
+class KafkaConsumerConfig(
+    @Value("\${kafka.bootstrap-servers}")
+    private val kafkaBootstrapServers: String,
+) {
+    @Bean
+    fun consumerConfigs(): Map<String, Any> =
+        mapOf(
+            ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG to kafkaBootstrapServers,
+            ConsumerConfig.GROUP_ID_CONFIG to "group_id",
+            ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG to StringDeserializer::class.java,
+            ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG to StringDeserializer::class.java,
+        )
+
+    @Bean
+    fun kafkaListenerContainerFactory(): ConcurrentKafkaListenerContainerFactory<String, String> {
+        val factory = ConcurrentKafkaListenerContainerFactory<String, String>()
+        factory.consumerFactory = DefaultKafkaConsumerFactory(consumerConfigs())
+        return factory
+    }
+}
