+++ shared-spring/src/main/kotlin/com/kohan/shared/spring/kafka/config/KafkaProducerConfig.kt
@@ -1,0 +1,30 @@
+package com.kohan.shared.spring.kafka.config
+
+import org.apache.kafka.clients.producer.ProducerConfig
+import org.apache.kafka.common.serialization.StringSerializer
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.kafka.core.DefaultKafkaProducerFactory
+import org.springframework.kafka.core.KafkaTemplate
+import org.springframework.kafka.core.ProducerFactory
+
+@Configuration
+class KafkaProducerConfig(
+    @Value("\${kafka.bootstrap-servers}")
+    private val kafkaBootstrapServers: String,
+) {
+    @Bean
+    fun producerConfigs(): Map<String, Any> =
+        mapOf(
+            ProducerConfig.BOOTSTRAP_SERVERS_CONFIG to kafkaBootstrapServers,
+            ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG to StringSerializer::class.java,
+            ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG to StringSerializer::class.java,
+        )
+
+    @Bean
+    fun producerFactory(): ProducerFactory<String, String> = DefaultKafkaProducerFactory(producerConfigs())
+
+    @Bean
+    fun kafkaTemplate(): KafkaTemplate<String, String> = KafkaTemplate(producerFactory())
+}
